Move 'bdd_from' and 'zdd_from' into their own compilation unit
