+++ src/adiar/bdd/convert.cpp
@@ -1,0 +1,36 @@
+#include <adiar/bdd.h>
+#include <adiar/bdd/bdd_policy.h>
+#include <adiar/zdd/zdd_policy.h>
+
+#include <adiar/domain.h>
+#include <adiar/internal/algorithms/convert.h>
+#include <adiar/internal/algorithms/intercut.h>
+
+namespace adiar
+{
+  __bdd bdd_from(const exec_policy &ep,
+                 const zdd &A,
+                 const generator<bdd::label_type> &dom)
+  {
+    return internal::intercut<internal::convert_dd_policy<bdd_policy, zdd_policy>>
+      (ep, A, dom);
+  }
+
+  __bdd bdd_from(const zdd &A, const generator<bdd::label_type> &dom)
+  {
+    return bdd_from(exec_policy(), A, dom);
+  }
+
+  __bdd bdd_from(const exec_policy &ep, const zdd &A)
+  {
+    const internal::shared_file<bdd::label_type> dom = domain_get();
+    internal::file_stream<domain_var> ds(dom);
+
+    return bdd_from(ep, A, make_generator(ds));
+  }
+
+  __bdd bdd_from(const zdd &A)
+  {
+    return bdd_from(exec_policy(), A);
+  }
+}
